WIP: arm64/kasan: support KASAN_VMALLOC
When KASAN_VMALLOC is selected, KASAN will lazily allocate shadow memory
for the vmalloc region, and we should not pre-populate this shadow
region with early (read-only zero) memory.

We need zero shadow for the fixmap, vmemmap, etc, so we must provide
early (read-only zero) shadow memory for everything between the end of
the vmalloc region and the start of the linear map.

TODO:

* Figure out the gap between VMALLOC_END and FIXADDR_START. It looks
  like these are forced into distinct PUDs, but it's not clear to me if
  anything else lives in this range.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>