arm64: enable KCSAN support
This patch allows KCSAN to be selected on arm64.

Prior to clang 12.0.0, this would interact poorly with BTI, producing
warnings for the majority of C files:

|   CC      arch/arm64/kernel/psci.o
| warning: some functions compiled with BTI and some compiled without BTI
| warning: not setting BTI in feature flags

Previously, boot issues have been seen where the alternatives patching
code was instrumented, and some rework is necessary to ensure that this
doesn't get instrumented in an unsound way.

TODO:
* Add a clang version dependency
* Check/ensure that patching code is not instrumented
* Fix races detected at boot time
* When did clang & GCC start respecting noinstr for tsan?
* Check instrumentation of patching code

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Marco Elver <elver@google.com>
Cc: Will Deacon <will@kernel.org>