arm64: entry: convert irq entry to C
Portions of our IRQ entry logic are handled in C while other parts are
handled in assembly. Let's migrate the bulk of it to C so that it's
easier to follow and maintain.

The original assembly called trace_hardirqs_off() for IRQs and NMIs, but
did not call trace_hardirqs_on() for NMIs where IRQs were priority
masked. The C code always balances these calls.

The now unused asm_nmi_{enter,exit}() wrappers are removed.

The EL0 BP hardening is moved from fault.c so that it's in the same
compilation unit as its only caller, where the compiler can inline it.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Will Deacon <will@kernel.org>