arm64: entry: convert error entry to C
Portions of our error entry logic are handled in C while other parts are
handled in assembly. Let's migrate the bulk of it to C so that it's
easier to follow and maintain.

The existing el*_error entry paths only unmasked debug exceptions, but
per the nesting rules introduced in commit:

  65be7a1b799f11ff ("arm64: introduce an order for exceptions")

... those should also unmask FIQ, as is the case with DAIF_ERRCTX.

Some now-unused macros are removed.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Will Deacon <will@kernel.org>