WIP: arm64: rework irq/daif flag manipulation
Over time, our DAIF and irqflag handling has become a nest of
edge-cases, making it very difficult to reason about.

To make things more reliable and easier to reason about, let's always
save/restore the DAIF flags, and additionally save/restore the PMR when
pNMI is in use. This ensures that save()...restore() sequences will
always behave as we expect, even during sequences where the PMR or DAIF
may be in an arbitrary state.

When restoring either irq or DAIF flags, flag tracing requires us to
look at both to determine whether IRQs will be enabled, so we
consistently save both even where we only need to adjust one of the two.

With this done, we no longer need to save the original DAIF.I state in
PMR, and can get rid of GIC_PRIO_PSR_I_SET and the corresponding
adjustments in the entry code.

As local_daif_unmask_procctx() now only needs to unmask DAIF,
local_daif_init_procctx() is built atop of this.

NOTE: This is *BROKEN* and does not interact well with patching of
static keys or alternatives. The init functions also assume a more
consistent environment than is present. This needs major rework.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Will Deacon <will@kernel.org>
Cc: Marc Zyngier <maz@kernel.org>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>