kvm: arm/arm64: save/restore DAIF in kvm_vcpu_run_vhe()
Future patches will change the way the local_daif_*() helpers work, and
it will only be valid to use local_daif_restore() when it is paired with
a call to local_daif_save().

In preparation for this, use a local_daif_save() .. local_daif_restore()
sequence in kvm_vcpu_run_vhe() rather than passing a synthetic value to
local_daif_restore().

As the flags should be equivalent to DAIF_PROCCTX_NOIRQ, there should be
no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>