arm64: sysreg: add DAIF (immediate) definitions
In subsequent patches we'll want to mess with DAIFSet and DAIFClr in
more places. Rather than open-coding the immediate values everywhere,
let's add common definitions and use these across arm64.

In subsequent patches we'll also want to use DAIFClr and DAIFSet in more
places, so helpers are added for manipulating these, and made use of
across arm64.

The open-coded instances in <asm/irqflags.h> are left as-is for now as
these will be substantially reworked in subsequent patches. Moving to
the new helpers means the assembly comment in local_daif_mask is
removed, but as this is only of use for debugging the assembly itself,
this is not a major concern.

There should be no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>