arm64: refactor DAIF+PMR unmasking
Subsequent patches will make it illegal to use local_daif_restore()
to initialize DAIF and PMR using a synthetic value. In preparation for
this, we need to replace such usage with helpers that perform equivalent
work.

This patch adds those helpers, and alters code to use them.

Entry code is mechanically converted to use local_daif_unmask_procctx()
and local_daif_unmask_procctx_noirq(), which are equivalent to to
contemporary usage of local_daif_restore(). Likewise in
do_notify_resume(), which happens on return from an EL0 exception, with
all exceptions masked in DAIF.

There are a few special cases to consider:

* In setup_arch() we use local_daif_restore(DAIF_PROCCTX_NOIRQ) today,
  and because this happens before ARM64_HAS_IRQ_PRIO_MASKING is
  detected, this always clears DA_F without touching the PMR. It is
  replaced with code which explicitly does just this, and the comment is
  updated.

* In init_IRQ() we use local_daif_restore(DAIF_PROCCTX_NOIRQ) to
  initialize the PMR into an expected state. This is given a helper to
  explicitly initalize the PMR as subsequent patches may not always do
  so in local_daif_unmask_procctx_noirq().

* In secondary_start_kernel() we use local_daif_restore(DAIF_PROCCTX) to
  initialize the PMR into an expected state. This is given a helper to
  explicitly initalize the PMR as subsequent patches may not always do
  so in local_daif_unmask_procctx_noirq().

There should be no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>