arm64: sys32: consistently use compat types
Currently, arm64's compat syscall wrappers don't consistently use compat
types, while uses of COMPAT_SYSCALL_DEFINE() in common code do.

This is not a functional problem today, but it would be preferable to
use compat types where the native type is larger than 32 bits, as this
will make it possible to automatically detect syscalls with missing
argument marshalling from the compat 32-bit PCS to the native 64-bit
PCS.

This patch updates our wrappers to use compat types. As int and unsigned
int are guranteed to be 32-bit regardless, these are left as-is,
matching the style of common code. Otherwise, these could be replaced
with compat_int_t and compat_uint_t respectively.

There should be no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Will Deacon <will@kernel.org>
Cc: Catalin Marinas <catalin.marinas@arm.com>