syscall: add NATIVE_SYSCALL_DEFINE*()
Some native syscalls cannot be used by compat code, and always require a
compat version to be defined. Let's make it possible to distinguish
these cases with new NATIVE_SYSCALL_DEFINE*() macros. This distinction
will allow for better compile-time sanity checks in subsequent patches.

As of this patch, a syscall can be define in three ways:

* NATIVE_SYSCALL_DEFINE() for a syscall that can only be used by native
  code.

* COMPAT_SYSCALL_DEFINE() for a syscall that can only be used by compat
  code.

* SYSCALL_DEFINE() for a syscall that can be used by both native and
  compat code.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>