arm64: Simplify __range_ok
The general case is not quite as compact as the inline assembly,
but with a sufficiently advanced compiler it is only 6 insns vs 5.

The real improvement comes from assuming that limit is never tiny,
and using __builtin_constant_p to make sure the constant folding
does not go awry.  This produces a 2 insn sequence even for older
compilers.

Signed-off-by: Richard Henderson <richard.henderson@linaro.org>
[Mark: rebase, use TASK_SIZE_MAX directly]
Signed-off-by: Mark Rutland <mark.rutland@arm.com>