bitfield.h: add helper to extract field width
To extract a signed bitfield, sign extension is required from the fields
most significant bit, and the (signed) target type is likely to differ
from the (unsigned) register value type.

To make it possible to handle such cases, add a helper to determine the
bit-width of a bitfield. The caller can then manually signed-extend an
extracted field to the desired type, e.g.

| u64 reg = SOME_VALUE;
| u64 val = FIELD_GET(reg, FIELDNAME);
| s64 sval = sign_extend64(val, FIELD_WIDTH(FIELDNAME));

Signed-off-by: Mark Rutland <mark.rutland@arm.com>