arm64: Kconfig: remove pNMI options
Presently pseudo NMI (pNMI) has no in-tree user, the structure of the
support code is getting in the way of substantial rework/cleanup
necessary for other features (e.g. RT), and there are a number of
potentially invasive changes that may need to be made to the support
code itself (e.g. to support systems regardless of FIQ routing).

While we would still like to make use of pNMI (e.g. to enable profiling
of regions with IRQs disabled, in the short term it would be preferable
to remove the existing support to enable other rework, such that we can
rebuild the pNMI support atop.

This patch removes arm64's Kconfig options for pNMI, and removes the
documentation for the now-unusable command line parameter. The support
code itself will be removed in subsequent patches.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Alexandru Elisei <alexandru.elisei@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>