arm64: convert error entry to C
Portions of our error entry logic are handled in C while other parts are
handled in assembly. Let's migrate the bulk of it to C so that it's
easier to follow and maintain.

The existing el*_error entry paths only unmasked debug exceptions, but
per the nesting rules introduced in commit:

  65be7a1b799f11ff ("arm64: introduce an order for exceptions")

... those should also unmask FIQ, as is the case with DAIF_ERRCTX.

On the el0 error return path we no longer unmask exceptions as there
will be immediately masked in ret_to_user.

The now-unsued enable_da_f asm macro is removed.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Alexandru Elisei <alexandru.elisei@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>