arm64: entry: wrap handler entry points
In subsequent patches we'll want to perform common entry/exit work when
transitioning to/from the entry assembly, with different entry/exit work
needed depending on the EL the exception was taken from.

It is tedious and error-prone to have each handler invoke this work
explicitly, so let's wrap these with a helper that does so, similar to
what DEFINE_SYSCALL() and friends do.

As ret_from_fork() returns to userspace without having entered the
kernel from an exception, this needs to manually explicitly invoke the
EL0 return work.

This patch only adds stubs for the entry/return work, which will be
expanded in subsequent patches.

There should be no functional change as a result of this patch.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Alexandru Elisei <alexandru.elisei@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: James Morse <james.morse@arm.com>
Cc: Julien Thierry <julien.thierry.kdev@gmail.com>
Cc: Marc Zyngier <maz@kernel.org>
Cc: Will Deacon <will@kernel.org>