arm64: alternatives: don't print while patching
In __apply_alternatives(), we log a message immediately before patching
the first alternative sequence we detect. While this is safe (as it
occurs before the first patch is applied), it's not generally safe to
call arbitrary code from __apply_alternatives(), and it'd be nicer to
factor this out to make it easier to verify that __apply_alternatives()
makes no unsafe calls.

This patch moves the pr_info_once() call out of __apply_alternatives()
and into apply_alternatives_all(). As this is only called once, we can
use pr_info() instead of pr_info_once(). This means we will always log
the message, even when there are no patches to apply, but as we apply
patches in the vast majority of cases this is not a significant loss.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Will Deacon <will@kernel.org>