arm64: mm: don't instrument physaddr functions
The __virt_to_phys() and __phys_addr_symbol() functions in physaddr.c
are used by lm_alias(), which is used in arm64's patching routines. As
instrumentation code may be patched (e.g. as it may use atomics) we must
ensure that these functions are not instrumented.

Mark the functions as `nosintr` to ensure this. The WARN() in
__virt_to_phys() may call patched code, but will only do so when things
are disastrously wrong, and we'd prefer a warning. The implementation of
WARN() transiently disables noinstr checks, so this shouldn't trip up
objtool in future.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Will Deacon <will@kernel.org>